refact-users: aprimoramentos nas validações de usuários, incluindo checagem de status ativo ao deletar e ajustes nas respostas de erro para operações de atualização e deleção

+++ users/src/modules/user2/application/use-cases/delete-user.usecase.ts
@@ -13,6 +13,10 @@ export class DeleteUserUseCase {
       throw new NotFoundItem('Não foi possivel encontrar o usuário');
     }
 
+    if (!user.active) {
+      throw new NotFoundItem('Não foi possivel encontrar o usuário');
+    }
+
     await this.userRepository.delete(id);
   }
 }
